Configure to ignore changes in Wok submodules.
After update or change any of the Wok submodules, the modifications
will be listed as Wok changes. Add configuration to ignore all changes
of in Wok submodules.

The value for each submodule can be "untracked", "dirty" or "all". When
"untracked" is used submodules are not considered dirty when they only
contain untracked content (but they are still scanned for modified
content). Using "dirty" ignores all changes to the work tree of
submodules, only changes to the commits stored in the superproject are
shown (this was the behavior before 1.7.0). Using "all" hides all
changes to submodules (and suppresses the output of submodule summaries
when the config option status.submodulesummary is set).

This patch sets "ignore = all" for all Wok submodules.

Signed-off-by: Paulo Vital <pvital@linux.vnet.ibm.com>