Upgrade objectstore schema.
With the new structure of the objectstore changed in Wok 2.0.0, the schema of
the objectstore file (when used by any plugin) needs to be checked and, if
necessary, updated to reflect the new structure.

This patch adds to utils.py file some common methods to support the operations
in the plugin's objectstore files. These methods were moved from Kimchi plugin
since they will be used by other plugins to do the same.

Signed-off-by: Paulo Vital <pvital@linux.vnet.ibm.com>