Issue #12: wokd start fails if plugin does not have ui/config/tab-ext.xml
Currently, it gets the list of tab-ext.xml
for all the installed plugins and then
each file is parsed in for loop.
Moved code which parse the tab-ext.xml file
inside try except block to handle IOError.
On IOError(File not found) continue the for
loop to parse other plugins tab-ext.xml.

Signed-off-by: Archana Singh <archus@linux.vnet.ibm.com>