utils.py: run_command runs in an en_US environment by default
This requirement started in a fixed Gingerbase bug (#18) that
reported lscpu problems when running in a French environment.
The output was being translated to french and breaking the
parsing made by the backend.

This is such a powder keg for all the plug-ins that I believe
this patch is justified. Instead of forcing en_US language
by using subprocess.Popen and setting the env, this is how
run_command will behave now:

- A new optional attribute, env_vars, was added. As the name
suggests, it allows the setup of the environment variables
to run the command.

- If env_vars is not set, run_command will copy the current
environment variables and set the language to the default
(en_US) by setting LC_ALL='C' variable. As of now, this is
the case of all the existing run_command calls in the code
for all WoK plug-ins. No behavior change will happen.

- If env_vars is set, but the LC_ALL var isn't, run_command
will set it to 'C' and force the default language.

- If env_vars is set and LC_ALL is also set, run_command
will not touch it and will run with env_vars as is. This
allows the caller to set the language at will.

Signed-off-by: Daniel Henrique Barboza <dhbarboza82@gmail.com>