Fixing loading of plug-ins without 'sub_nodes'
Recent WoK changes were made considering that all WoK plug-ins
have the control modules in a dir named 'control', but
Ginger, for historical/unknown reasons, uses the dir
'controls'. The result is that WoK became unable of loading
Ginger because the code was executing a 'continue' in the
'for' loop, preventing the loading of plug-in at all.

This patch amends it.

Signed-off-by: Daniel Henrique Barboza <dhbarboza82@gmail.com>