Added new Wok logo and script to display installed plugins icons and versions
This commit replaces the old icons, images and favicons with the new Wok logo.
It also includes a new script to generate and append the installed plugins logos
(when available) in the Login page, wok-ui footer and About window.
In order to add a new logo, a .svg or .png file with the plugin name must be
installed in each plugin /ui/images/ folder.

Signed-off-by: Samuel Guimarães <sguimaraes943@gmail.com>