Fixing WoK match in search queries
Prior to this patch, supposing an example with the VMs API
of Kimchi, we could send a GET request like this:

/plugins/kimchi/vms?state=thisisnotrunning

and get as a answer all the VMs with state=running. The reason
was that WoK was doing a membership verification in the input
value even if the value isn't a list and, in this case,
'running' in 'thisisnotrunning' was returning True  and the
VM would be added to the returned results.

This also impacted regular expressions with negative lookahead.
Using the same API mentioned above, a search query like
'state=^(?!running).*$' would return all VMs because if
state was not 'running' it would match this regex and, if
it was 'running', it would match this membership rule because
'running' in '^(?!running).*$' returns True.

Signed-off-by: Daniel Henrique Barboza <dhbarboza82@gmail.com>