Use system's nginx proxy service.
This patch removes the code that executes a dedicated nginx proxy, making Wok
to use the system's nginx service. This is a requirement to make Wok acceptable
in community repositories.

It also make sure that a Wok executed from path different than installed (from
a cloned and builded source code, for example) will create a symbolic link in
system's nginx config dir to the running configuration file.

This patch solves part of issue #25

Signed-off-by: Paulo Vital <pvital@linux.vnet.ibm.com>