Issue #122 - Make AsyncTask stoppable.
This patch gives to user a way to 'stop' a Task that is still running by setting
the Task status to "killed".

Since an AsyncTask is basic a thread running in the system and this thread can
execute a pure Python method or a background command (by using run_command()
from wok.utils), the developer must pass to AsyncTask constructor a method to be
executed by the DELETE operation, called here as 'kill_cb'. If none kill_cb is
passed, the task will not be able to stopped and an error message will be raised
to user if DELETE operation is executed. Otherwise, the kill_cb method will be
executed by kill() method (responsible to execute the DELETE operation) of
AsyncTask class and its status set to 'killed'.

Signed-off-by: Paulo Vital <pvital@linux.vnet.ibm.com>