Bug fix: Set application options globally on Server() instead of wokd command
wokd is a command line tool which relies on Server() instance to start
up the web server. To run tests, the wokd command is not used. Instead
of that, the Server() instance handles the application options.
So instead of setting the options globally on wokd command, do that on
Server() to make it used overwhere.

This bug was introcuded by 713fd5caa and noticed while running Kimchi
tests.

Signed-off-by: Aline Manera <alinefm@linux.vnet.ibm.com>