Use a CDN to get dataTables JS and CSS files
That way, Wok does not need to import/package the dataTables JS and CSS files.
Reducing the imported JS code might be easier to get the Wok package
acceptable on Debian and/or Fedora official repositories.

There is no need to keep a scss file for custom Wok configuration. Move
only what is needed to wok.scss and use the CDN for
dataTables.bootstrap.min.css for default style.

Signed-off-by: Aline Manera <aline.manera@gmail.com>